feat: Repository 주키 속성 및 커스텀 함수 작성

+++ src/main/java/com/yoen/yoen_back/repository/image/TravelRecordImageRepository.java
@@ -3,5 +3,5 @@ package com.yoen.yoen_back.repository.image;
 import com.yoen.yoen_back.entity.image.TravelRecordImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TravelRecordImageRepository extends JpaRepository<TravelRecordImage, Integer> {
+public interface TravelRecordImageRepository extends JpaRepository<TravelRecordImage, Long> {
 }
