+++ src/main/java/com/yoen/yoen_back/repository/payment/SettlementRepository.java
@@ -3,5 +3,5 @@ package com.yoen.yoen_back.repository.payment;
 import com.yoen.yoen_back.entity.payment.Settlement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SettlementRepository extends JpaRepository<Settlement, Integer> {
+public interface SettlementRepository extends JpaRepository<Settlement, Long> {
 }
